Renamed the files to match the names of the buttons and adapted the code accordingly
