--- createEvent/createEvent.html
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Create Event</title>
-    <link rel="stylesheet" href="createEventStyle.css">
-</head>
-<body>
-    
-
-    <header>
-        <section id="profileInfo">
-            <!-- take the information from a database about users -->
-            <img id="profilePicture" alt = "N/A"></img>
-            <p id="profileName">Profile name</p>
-        </section>
-        <nav id="topNav">
-            <ul>
-                <!-- make admin settings and create event appear only if the user has the rights to it -->
-                <li><a href="./adminSettings/adminSettings.html" id="adminSettings">Administrative Settings</a></li>
-                <li><a href="createEvent.html" id="createEvent">Create Event</a></li>
-                <li><a href="profile.html" id="profile">Profile</a></li>
-                <li><a href="login.html" id="logout">Log out</a></li>
-            </ul>
-        </nav>
-    </header>
-	
-	<form id="eventForm" action="createEvent.php" method="post">
-	   <label for="eventName">Event name:</label>
-	   <input type="text" id="eventName" name="eventName"><br><br>
-	   <label for="eventImg">Select image:</label>
-	   <input type="file" id="eventImg" name="eventImg" accept="image/*"><br><br>
-	   <label for="usersNum">Event capacity:</label>
-	   <input type="number" id="usersNum" name="usersNum" min="1" max="32"><br><br>
-	   <label for="eventDesc">Description:</label>
-	   <input type="text" id="eventDesc" name="eventDesc"><br><br>
-	   <input type="submit">
-	</form>
-
-    
-
-</body>
-</html>
